Cambio interno e la informacion
Ahora se alamcena en objetos

+++ DataLayer/LogicalLayer/Interfaces/MoneyManager.cs
@@ -6,6 +6,6 @@ namespace LogicalLayer
 {
     public interface MoneyManager
     {
-        void addMoney(int value);
+        void addMoney(float value);
     }
 }
